+++ src/tests/ui/test_ui_base.py
@@ -326,7 +326,11 @@ class Test(unittest.TestCase):
         https://github.com/AndreMiras/PyWallet/issues/90
         """
         controller = app.controller
+        pywalib = controller.pywalib
         manage_existing = controller.manage_existing
+        # makes sure an account is selected
+        pywalib.new_account(password="password", security_ratio=1)
+        controller.current_account = pywalib.get_account_list()[0]
         # ManageExisting and Controller current_account should be in sync
         self.assertEqual(
             manage_existing.current_account, controller.current_account)
@@ -344,6 +348,66 @@ class Test(unittest.TestCase):
         self.assertEqual(dialog.title, 'No account selected.')
         controller.dismiss_all_dialogs()
 
+    def helper_confirm_account_deletion(self, app):
+        """
+        Helper method for confirming account deletion popups.
+        """
+        controller = app.controller
+        manage_existing = controller.manage_existing
+        # a confirmation popup should show
+        dialogs = controller.dialogs
+        self.assertEqual(len(dialogs), 1)
+        dialog = dialogs[0]
+        self.assertEqual(dialog.title, 'Are you sure?')
+        # confirm it
+        # TODO: click on the dialog action button itself
+        manage_existing.on_delete_account_yes(dialog)
+        # the dialog should be replaced by another one
+        dialogs = controller.dialogs
+        self.assertEqual(len(dialogs), 1)
+        dialog = dialogs[0]
+        self.assertEqual(dialog.title, 'Account deleted, redirecting...')
+        controller.dismiss_all_dialogs()
+        self.assertEqual(len(dialogs), 0)
+
+    def helper_test_delete_account_twice(self, app):
+        """
+        Trying to delete the same account twice, shoult not crash the app:
+        https://github.com/AndreMiras/PyWallet/issues/51
+        """
+        controller = app.controller
+        pywalib = controller.pywalib
+        manage_existing = controller.manage_existing
+        # makes sure an account is selected
+        pywalib.new_account(password="password", security_ratio=1)
+        controller.current_account = pywalib.get_account_list()[0]
+        self.assertTrue(manage_existing.current_account is not None)
+        account_count_before = len(pywalib.get_account_list())
+        # let's try to delete this account once
+        delete_button_id = manage_existing.ids.delete_button_id
+        delete_button_id.dispatch('on_release')
+        self.helper_confirm_account_deletion(app)
+        # the account should be deleted
+        self.assertEqual(
+            len(pywalib.get_account_list()), account_count_before - 1)
+        # makes sure the account was also cleared from the selection view
+        switch_account = self.helper_load_switch_account(app)
+        account_list_id = switch_account.ids.account_list_id
+        self.assertEqual(
+            len(account_list_id.children), len(pywalib.get_account_list()))
+        # TODO: the selected account should now be None
+        # self.assertIsNone(manage_existing.current_account)
+        # self.assertIsNone(controller.current_account)
+        # let's try to delete this account a second time
+        delete_button_id = manage_existing.ids.delete_button_id
+        delete_button_id.dispatch('on_release')
+        # TODO: the second time an error dialog should pop
+        # dialogs = controller.dialogs
+        # self.assertEqual(len(dialogs), 1)
+        # dialog = dialogs[0]
+        # self.assertEqual(dialog.title, 'No account selected.')
+        controller.dismiss_all_dialogs()
+
     # main test function
     def run_test(self, app, *args):
         Clock.schedule_interval(self.pause, 0.000001)
@@ -354,6 +418,7 @@ class Test(unittest.TestCase):
         self.helper_test_address_alias(app)
         self.helper_test_delete_account(app)
         self.helper_test_delete_account_none_selected(app)
+        self.helper_test_delete_account_twice(app)
 
         # Comment out if you are editing the test, it'll leave the
         # Window opened.